Use push256 for outer tag on APT, separate 'AC' tags per algo
This corrects the malformed BER-TLV being generated on some
builds (when strings in the APT are long enough), and makes
our APT a little closer to standards-compliant.

To get back a little of the space we're paying for all the
'AC' tags per algo, we'll stop pushing quite so many into
the list (just one AES, for example).

See #45